+++ src/models/user.ts
@@ -19,10 +19,6 @@ const userSchema = new Schema(
             required: true,
             trim: true,
         },
-        viewed: {
-            type: [String],
-            required: true,
-        },
         deviceInfo: {
             fingerprint: String,
             brand: String,